requirements.txt
flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.2
Werkzeug==2.3.7
itsdangerous==2.2.0
click==8.1.7
blinker==1.6.2
